Moved flag setters to new source file, and add detailed docstrings.
